requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
Pillow==8.3.2
psycopg2-binary==2.9.1
Flask-JWT-Extended==4.4.4
PyJWT==2.6.0
bcrypt==4.0.1
